Adding a controller abstraction layer
